refactor: refactor spinner tests and add accessible role to the spinner
